Crate csshw_lib

Cluster SSH tool for Windows inspired by csshX

Modules

cli

CLI interface

client

Client implementation

daemon

Daemon imlementation

serde

Serialization/Deserialization implemention for windows INPUT_RECORD_0.

utils

Utilities shared by daemon and client.

Structs

WindowsSettingsDefaultTerminalApplicationGuard

Guard storing previous/old DelegationConsole and DelegationTerminal registry values.

Constants

CLSID_CONHOST 🔒

CLSID identifying conhost.exe in the registry.

CLSID_DEFAULT 🔒

CLSID identifying the default configuration in the registry.

DEFAULT_TERMINAL_APP_REGISTRY_PATH 🔒

Registry path where DelegationConsole and DelegationTerminal registry keys are stored.

DELEGATION_CONSOLE 🔒

DelegationConsole registry key.

DELEGATION_TERMINAL 🔒

DelegationTerminal registry key.

Functions

enumerate_callback 🔒 ⚠

Callback function used in enumerate_windows to pass a Rust closure to windows C code.

enumerate_windows 🔒

Enumerate all top-level windows on the screen and call the given callback for each.

get_concole_window_handle

Return the Window Handle [HWND] for the foreground window associated with the given process_id.

get_reg_key 🔒

Retrieve the [RegKey] under which the default terminal application system settings are stored.

get_registry_value 🔒

Try to retrieve the registry value for the given value_name from the given regkey.

init_logger

Initialize the logger.

spawn_console_process

Launch the given console application with the given arguments as a new detached process with its own console window.

write_registry_values 🔒

Write DelegationConsole and DelegationTerminal registry values to the given [RegKey].