csshw_lib::cli
Function resolve_cluster_tags

fn resolve_cluster_tags<'a>(
    hosts: Vec<&'a str>,
    clusters: &'a Vec<Cluster>,
) -> Vec<&'a str>

Resolve cluster tags into hostnames

Iterates over the list of hosts to find and resolve cluster tags. Nested cluster tags are supported but recursivness is not checked for.

Arguments

hosts - List of hosts including hostnames and or cluster tags

clusters - List of available cluster tags

Returns

A list of hostnames