csshw_lib::daemon
Function arrage_client_window

fn arrage_client_window(
    handle: &HWND,
    workspace_area: &WorkspaceArea,
    index: usize,
    number_of_consoles: usize,
    aspect_ratio_adjustment: f64,
)

Re-sizes and re-positions the given client window based on the total number of clients, the size of the daemon console window and its index relative to the other client windows.

Arguments

handle - Reference the windows handle of a client console window.

workspace_area - The available workspace area on the primary monitor minus the space occupied by the daemon console window.

index - The index of the client in the list of all clients.

number_of_consoles - The total number of active client console windows.

aspect_ration_adjustment - The aspect_ratio_adjustment daemon configuration.