csshw_lib::daemon
Function launch_clients

async fn launch_clients(
    hosts: Vec<String>,
    username: &Option<String>,
    debug: bool,
    workspace_area: &WorkspaceArea,
    aspect_ratio_adjustment: f64,
    index_offset: usize,
) -> BTreeMap<usize, ClientWindow>

Launches a client console for each given host and waits for the client windows to exist before returning their handles.

Arguments

hosts - List of hosts

username - Optional username, if none is given the client will use the SSH config to determine a username.

debug - Toggles debug mode on the client.

workspace_area - The available workspace area on the primary monitor minus the space occupied by the daemon console window. Used to arrange the client window.

aspect_ratio_adjustment - The aspect_ratio_adjustment daemon configuration. Used to arrange the client window.

index_offset - Offset used to position the new windows correctly from the start, avoiding flickering.

Returns

A mapping from the order a client console window was launched at in relation to the other client windows and the clients console window handle.