csshw_lib::daemon
Function main

pub async fn main(
    hosts: Vec<String>,
    username: Option<String>,
    config: &DaemonConfig,
    debug: bool,
)

The entrypoint for the daemon subcommand.

Spawns 1 client process with its own window for each host and 1 worker thread that handles communication with the client over a named pipe. Responsible for client window positioning and sizing. Handles control mode. Main thread reads input records from the console input buffer and propagates them via the background threads to all clients simultaneously.

Arguments

hosts - List of hostnames for which to launch clients.

username - Username used to connect to the hosts. If none, each client will use the SSH config to determine a suitable username for their respective host.

config - The DaemonConfig.

debug - Enables debug logging