csshw_lib::daemon
Module workspace

Structs

WorkspaceArea

The available workspace area on the primary monitor

Enums

Scaling

Possible scalings for any pixel related values

Functions

get_primary_monitor 🔒

Returns a handle to the primary monitor.

get_scale_factor 🔒

Returns the scaling factor of the primary monitor.

get_workspace_area

Returns the available workspace area on the primary monitor in the specified scaling minus the space occupied by the daemon console.