csshw_lib
Module utils

Utilities shared by daemon and client.

Modules

config

Client and Daemon configuration structs.

constants

Shared constants.

debug

Implements string representations for INPUT_RECORD related structs. For debugging only.

Constants

KEY_EVENT 🔒

u16 representation of a KEY_EVENT.

Functions

arrange_console

Changes size and position of the current console window.

clear_screen

Empties the console screen output buffer of the current console window.

get_console_input_buffer

Returns a [HANDLE] to the [STD_INPUT_HANDLE] of the current process.

get_console_output_buffer

Returns a [HANDLE] to the [STD_OUTPUT_HANDLE] of the current process.

get_console_title

Returns the title of the current console window.

get_std_handle 🔒

Returns a [HANDLE] to the requested [STD_HANDLE] of the current process.

get_window_title

Returns the title of the window represented by the given window handle [HWND].

is_windows_10

Detects if the current windows installation is Windows 10 or not.

print_console_rect

Continously prints the window rectangle of the current console window.

read_console_input 🔒

Returns a single [INPUT_RECORD] read from the current process stdinput.

read_keyboard_input

Returns a single [INPUT_RECORD_0] where EventType == KEY_EVENT.

set_console_border_color

Sets the border color of the current console window.

set_console_color

Sets the back- and foreground color of the current console window.

set_console_title

Sets the window title of the current console window.